// src/api/endpoints/api.schemas.ts
export interface ObjectPerms {
  objectId?: string;
  /** @nullable */
  userId: string | null;
  renovationOnly?: boolean;
  canEdit?: boolean;
}